+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Improvement: Reactify admin pages (Customize)
 * Improvement: Ensure not to consider `[text|site](https://example.com]` as a row in the table
+* Improvement: Enter key behavior in markdown table
 * Fix: Emoji Autocomplete window does not float correctly
     * Introduced by 3.5.0